+++ sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/plans/logical/v2Commands.scala
@@ -32,6 +32,7 @@ import org.apache.spark.sql.types.{DataType, MetadataBuilder, StringType, Struct
 trait V2WriteCommand extends Command {
   def table: NamedRelation
   def query: LogicalPlan
+  def isByName: Boolean
 
   override def children: Seq[LogicalPlan] = Seq(query)
 
@@ -50,6 +51,8 @@ trait V2WriteCommand extends Command {
             (outAttr.nullable || !inAttr.nullable)
       })
   }
+
+  def withNewQuery(newQuery: LogicalPlan): V2WriteCommand
 }
 
 /**
@@ -59,7 +62,9 @@ case class AppendData(
     table: NamedRelation,
     query: LogicalPlan,
     writeOptions: Map[String, String],
-    isByName: Boolean) extends V2WriteCommand
+    isByName: Boolean) extends V2WriteCommand {
+  override def withNewQuery(newQuery: LogicalPlan): AppendData = copy(query = newQuery)
+}
 
 object AppendData {
   def byName(
@@ -89,6 +94,9 @@ case class OverwriteByExpression(
   override lazy val resolved: Boolean = {
     table.resolved && query.resolved && outputResolved && deleteExpr.resolved
   }
+  override def withNewQuery(newQuery: LogicalPlan): OverwriteByExpression = {
+    copy(query = newQuery)
+  }
 }
 
 object OverwriteByExpression {
@@ -116,7 +124,11 @@ case class OverwritePartitionsDynamic(
     table: NamedRelation,
     query: LogicalPlan,
     writeOptions: Map[String, String],
-    isByName: Boolean) extends V2WriteCommand
+    isByName: Boolean) extends V2WriteCommand {
+  override def withNewQuery(newQuery: LogicalPlan): OverwritePartitionsDynamic = {
+    copy(query = newQuery)
+  }
+}
 
 object OverwritePartitionsDynamic {
   def byName(
